# Add support for container pooling