--- a/roaring/containers.go
+++ b/roaring/containers.go
@@ -19,10 +19,124 @@ type sliceContainers struct {
 	containers    []*Container
 	lastKey       uint64
 	lastContainer *Container
+
+	containersPool containersPool
+}
+
+// ContainerPoolingConfiguration represents the configuration for
+// container pooling.
+type ContainerPoolingConfiguration struct {
+	// Whether an array should be allocated for each pooled container.
+	AllocateArray bool
+	// Whether a bitmap should be allocated for each pooled container.
+	AllocateBitmap bool
+	// Whether a run should be allocated for each pooled container.
+	AllocateRuns bool
+
+	// Maximum number of containers to pool.
+	MaxCapacity int
+	// Maximum size of keys and containers slice to maintain after calls to Reset().
+	MaxKeysAndContainersSliceLength int
+}
+
+// NewDefaultContainerPoolingConfiguration creates a ContainerPoolingConfiguration
+// with default configuration.
+func NewDefaultContainerPoolingConfiguration(maxCapacity int) ContainerPoolingConfiguration {
+	return ContainerPoolingConfiguration{
+		AllocateArray:  true,
+		AllocateRuns:   true,
+		AllocateBitmap: true,
+
+		MaxCapacity: maxCapacity,
+
+		// Compared to the size of the containers themselves, these slices are
+		// very small (8 bytes per entry), so we can afford to allow them to
+		// grow larger.
+		MaxKeysAndContainersSliceLength: maxCapacity * 10,
+	}
+}
+
+type containersPool struct {
+	containers []*Container // Nil if pooling disabled, otherwise present.
+	config     ContainerPoolingConfiguration
+}
+
+func (cp *containersPool) put(c *Container) {
+	if cp == nil || cp.containers == nil {
+		// Ignore if pooling isn't configured.
+		return
+	}
+
+	if c.mapped {
+		// If the container was mapped, assume the whole thing has been
+		// corrupted and reset it to an initial state (without reallocating
+		// according to the pooling config as its likely it will just be
+		// mapped again.)
+		*c = newContainer()
+		cp.containers = append(cp.containers, c)
+		return
+	}
+
+	if len(cp.containers) >= cp.config.MaxCapacity {
+		// Don't allow pool to exceed maximum capacity.
+		return
+	}
+
+	if c.array != nil && !cp.config.AllocateArray {
+		// Don't allow any containers with an allocated array slice to be
+		// returned to the pool if the config doesn't allow it.
+		return
+	}
+
+	if c.runs != nil && !cp.config.AllocateRuns {
+		// Don't allow any containers with an allocated run slice to be
+		// returned to the pool if the config doesn't allow it.
+		return
+	}
+
+	// Reset before returning to the pool to ensure all calls to get() return
+	// a clean container.
+	c.Reset()
+	cp.containers = append(cp.containers, c)
+}
+
+func (cp *containersPool) get() *Container {
+	if cp == nil || cp.containers == nil {
+		return NewContainer()
+	}
+
+	if len(cp.containers) > 0 {
+		// If we have a pooled container available, use that.
+		lastIdx := len(cp.containers) - 1
+		c := cp.containers[lastIdx]
+		cp.containers = cp.containers[:lastIdx]
+		return c
+	}
+
+	// Pooling is enabled, but there are no available containers,
+	// so we allocate.
+	return NewContainerWithPooling(cp.config)
 }
 
 func newSliceContainers() *sliceContainers {
 	return &sliceContainers{}
+}
+
+func newSliceContainersWithPooling(poolingConfig ContainerPoolingConfiguration) *sliceContainers {
+	sc := &sliceContainers{
+		keys:       make([]uint64, 0, poolingConfig.MaxCapacity),
+		containers: make([]*Container, 0, poolingConfig.MaxCapacity),
+	}
+
+	sc.containersPool = containersPool{
+		config:     poolingConfig,
+		containers: make([]*Container, 0, poolingConfig.MaxCapacity),
+	}
+	for i := 0; i < poolingConfig.MaxCapacity; i++ {
+		sc.containersPool.put(NewContainerWithPooling(poolingConfig))
+	}
+
+	return sc
 }
 
 func (sc *sliceContainers) Get(key uint64) *Container {
@@ -49,7 +163,7 @@ func (sc *sliceContainers) Put(key uint64, c *Container) {
 func (sc *sliceContainers) PutContainerValues(key uint64, containerType byte, n int, mapped bool) {
 	i := search64(sc.keys, key)
 	if i < 0 {
-		c := NewContainer()
+		c := sc.containersPool.get()
 		c.containerType = containerType
 		c.n = int32(n)
 		c.mapped = mapped
@@ -69,7 +183,9 @@ func (sc *sliceContainers) Remove(key uint64) {
 	if i < 0 {
 		return
 	}
+
 	sc.keys = append(sc.keys[:i], sc.keys[i+1:]...)
+	sc.containersPool.put(sc.containers[i])
 	sc.containers = append(sc.containers[:i], sc.containers[i+1:]...)
 
 }
@@ -93,7 +209,7 @@ func (sc *sliceContainers) GetOrCreate(key uint64) *Container {
 	sc.lastKey = key
 	i := search64(sc.keys, key)
 	if i < 0 {
-		c := NewContainer()
+		c := sc.containersPool.get()
 		sc.insertAt(key, c, -i-1)
 		sc.lastContainer = c
 		return c
@@ -104,11 +220,25 @@ func (sc *sliceContainers) GetOrCreate(key uint64) *Container {
 }
 
 func (sc *sliceContainers) Clone() Containers {
-	other := newSliceContainers()
-	other.keys = make([]uint64, len(sc.keys))
+	var other *sliceContainers
+	if sc.containersPool.containers != nil {
+		other = newSliceContainersWithPooling(sc.containersPool.config)
+	} else {
+		other = newSliceContainers()
+	}
+
+	if cap(other.keys) > len(sc.keys) {
+		other.keys = other.keys[:len(sc.keys)]
+	} else {
+		other.keys = make([]uint64, len(sc.keys))
+	}
+
 	other.containers = make([]*Container, len(sc.containers))
 	copy(other.keys, sc.keys)
 	for i, c := range sc.containers {
+		// TODO(rartoul): It would be more efficient to use one of
+		// other's pooled containers instead of alllowing Clone to
+		// allocate a new one when pooling is enabled.
 		other.containers[i] = c.Clone()
 	}
 	return other
@@ -135,10 +265,36 @@ func (sc *sliceContainers) Count() uint64 {
 }
 
 func (sc *sliceContainers) Reset() {
-	sc.keys = sc.keys[:0]
-	sc.containers = sc.containers[:0]
+	for i := range sc.containers {
+		// Try and return containers to the pool (no-op if disabled.)
+		sc.containersPool.put(sc.containers[i])
+		// Clear pointers to allow G.C to reclaim objects if these were the
+		// only outstanding pointers.
+		sc.containers[i] = nil
+	}
+
+	if sc.poolingEnabled() {
+		if cap(sc.keys) <= sc.containersPool.config.MaxKeysAndContainersSliceLength {
+			sc.keys = sc.keys[:0]
+		} else {
+			sc.keys = make([]uint64, 0, sc.containersPool.config.MaxCapacity)
+		}
+
+		if cap(sc.containers) <= sc.containersPool.config.MaxKeysAndContainersSliceLength {
+			sc.containers = sc.containers[:0]
+		} else {
+			sc.containers = make([]*Container, 0, sc.containersPool.config.MaxCapacity)
+		}
+	} else {
+		sc.keys = sc.keys[:0]
+		sc.containers = sc.containers[:0]
+	}
 	sc.lastContainer = nil
 	sc.lastKey = 0
+}
+
+func (sc *sliceContainers) poolingEnabled() bool {
+	return sc.containersPool.containers != nil
 }
 
 func (sc *sliceContainers) seek(key uint64) (int, bool) {