--- a/roaring/roaring.go
+++ b/roaring/roaring.go
@@ -1354,16 +1354,16 @@ func NewContainerWithPooling(poolingConfig ContainerPoolingConfiguration) *Conta
 		containerType: containerArray,
 	}
 
-	if poolingConfig.AllocateArray {
-		c.array = make([]uint16, 0, ArrayMaxSize)
+	if poolingConfig.MaxArraySize > 0 {
+		c.array = make([]uint16, 0, poolingConfig.MaxArraySize)
 	}
 	if poolingConfig.AllocateBitmap {
 		c.bitmap = make([]uint64, bitmapN)
 		// Start as a bitmap since we had to allocate it anyways.
 		c.containerType = containerBitmap
 	}
-	if poolingConfig.AllocateRuns {
-		c.runs = make([]interval16, 0, runMaxSize)
+	if poolingConfig.MaxRunsSize > 0 {
+		c.runs = make([]interval16, 0, poolingConfig.MaxRunsSize)
 	}
 
 	return c