… roaring/roaring_internal_test.go