## Overview

This pull request adds support for pooling containers as discussed in #1765

The idea is that the existing implementation is wasteful when bitmaps are short-lived because their is no way to re-use them (and their containers). This P.R adds the ability to pre-allocate containers on a given Bitmap, as well as Reset() a bitmap so that those contains can be reused. This allows the bitmaps to be pooled at the application level and can heavily reduce allocations and G.C pressure in situations where bitmaps are short lived.

Before
```
goos: darwin
goarch: amd64
pkg: github.com/pilosa/pilosa/roaring
BenchmarkUnionBulk-8   	  100000	     17075 ns/op	   33536 B/op	      21 allocs/op
PASS
ok  	github.com/pilosa/pilosa/roaring	1.902s
Success: Benchmarks passed.
``` 

After
```
goos: darwin
goarch: amd64
pkg: github.com/pilosa/pilosa/roaring
BenchmarkUnionBulkWithPooling-8   	  200000	     11738 ns/op	     172 B/op	       5 allocs/op
PASS
ok  	github.com/pilosa/pilosa/roaring	2.494s
Success: Benchmarks passed.
```

Fixes #1765

## Pull request checklist

- [X] I have read the [contributing guide](https://github.com/pilosa/pilosa/blob/master/CONTRIBUTING.md).
- [X] I have agreed to the [Contributor License Agreement](https://cla-assistant.io/pilosa/pilosa).
- [X] I have updated the [documentation](https://github.com/pilosa/pilosa/tree/master/docs).
- [X] I have resolved any merge conflicts.
- [x] I have included tests that cover my changes.
- [x] All new and existing tests pass.

## Code review checklist
This is the checklist that the reviewer will follow while reviewing your pull request. You do not need to do anything with this checklist, but be aware of what the reviewer will be looking for.

- [ ] Ensure that any changes to external docs have been included in this pull request.
- [ ] If the changes require that minor/major versions need to be updated, tag the PR appropriately.
- [ ] Ensure the new code is [properly commented](https://github.com/golang/go/wiki/CodeReviewComments#doc-comments) and follows [Idiomatic Go](https://dmitri.shuralyov.com/idiomatic-go).
- [ ] Check that tests have been written and that they cover the new functionality.
- [ ] Run tests and ensure they pass.
- [ ] Build and run the code, performing any applicable integration testing.