@jaffee 2019-06-11
we've diverged from this, and there are some coming changes to sync.Pool in Go 1.13... we will likely revisit something like this in the future.